# server/src/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lingshan AI Guide API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat/ask": {
      "post": {
        "summary": "RAG guide answer: grounded reply, citations, quality/emotion label and latency breakdown",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": { "type": "string" },
                  "sessionId": { "type": "string" },
                  "mode": { "type": "string", "enum": ["qa", "explain", "route"] },
                  "history": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": { "type": "string", "enum": ["user", "assistant"] },
                        "content": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}